Add --static compiler option
When attempting to statically link the compiler, we need to statically link
llvm. This means we need to pass --link-static to llvm-config. This is achieved
by adding an official --static compiler option, which both adds -static to the
cc invocation, and sets a "static" compiler flag. This is utilised in
lib_llvm.cr to add --link-static to our llvm-config invocation.